pyproject.toml
[tool.poetry]
name = "productapi"
version = "0.1.0"
description = "A Product API for Summy Store"
authors = ["Hardeep Kumar <hardeepkumar1815@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.1"
yapf = "^0.32.0"
djangorestframework = "^3.13.1"
django-jazzmin = "^2.4.9"
python-dotenv = "^0.19.2"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
dj-database-url = "^0.5.0"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
